refactor : URI 명세 수정
- /members -> /admin/members
