+++ src/main/java/roomescape/admin/service/AdminService.java
@@ -58,4 +58,8 @@ public class AdminService {
 
         return searchedReservations;
     }
+
+    public List<Member> findAll() {
+        return memberRepository.findAll();
+    }
 }
