+++ src/main/java/roomescape/member/service/MemberService.java
@@ -1,6 +1,5 @@
 package roomescape.member.service;
 
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.common.exception.DataNotFoundException;
@@ -16,9 +15,5 @@ public class MemberService {
     public Member findMemberByEmail(final String email) {
         return memberRepository.findByEmail(email)
                 .orElseThrow(() -> new DataNotFoundException("해당 회원 데이터가 존재하지 않습니다. email = " + email));
-    }
-
-    public List<Member> findAll() {
-        return memberRepository.findAll();
     }
 }
