+++ src/test/java/roomescape/member/controller/MemberRestControllerTest.java
@@ -1,6 +1,8 @@
 package roomescape.member.controller;
 
 import io.restassured.RestAssured;
+import io.restassured.http.ContentType;
+import java.util.Map;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
@@ -15,8 +17,16 @@ public class MemberRestControllerTest {
 
     @Test
     void 회원_목록을_조회한다() {
+        //given
+        final String adminToken = RestAssured.given()
+                .contentType(ContentType.JSON)
+                .body(Map.of("email", "east@email.com", "password", "1234"))
+                .when().post("/login").getCookie("token");
+
+        //wben & then
         RestAssured.given().log().all()
-                .when().get("/members")
+                .cookie("token", adminToken)
+                .when().get("/admin/members")
                 .then().log().all()
                 .statusCode(HttpStatus.OK.value())
                 .body("size()", is(2));
